% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files-validation.R
\name{create_release_files}
\alias{create_release_files}
\title{Create release files}
\usage{
create_release_files(OSF_LINK, OSF_TOKEN, dictionary_path = NULL)
}
\arguments{
\item{OSF_LINK}{link to the dictionary stored on OSF used for updating.}

\item{OSF_TOKEN}{OSF auth token used for modifying OSF directories}

\item{dictionary_path}{optional string containing path to the dictionary directory. When provided no dictionary is downloaded.}
}
\description{
Creates release files given the user OSF link and auth token.
}
